+++ src/pages/Payment/Qrcode.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useEffect } from "react"
 import { useLocation, useNavigate, useParams } from "react-router-dom"
 import { FloatingGlassBox } from "../../components/Containers"
-import { formatTime, useTimer } from "../../components/Timer/Timer"
+import { formatTime, useTimer } from "../../hooks/useTimer"
 import { PaymentApiResponse } from "../../utils/types/shop"
 
 export const QrcodePage = () => {
@@ -21,10 +21,10 @@ export const QrcodePage = () => {
   }, [paymentState?.download_uri, startTimer, navigate])
   return (
     <FloatingGlassBox>
-      <div className=" flex flex-col justify-center py-8 px-8">
+      <div className=" flex flex-col justify-center py-8 px-4">
         <img src={paymentState?.download_uri || "https://picsum.photos/283/328"} />
         <div className="mt-5 flex flex-col items-center text-white">
-          <h1 className="text-xl font-bold sm:text-3xl">ยอดเงินของรายการสั่งซื้อนี้ : {paymentState?.amount}</h1>
+          <h1 className="text-xl font-bold sm:text-3xl">ยอดเงินของรายการสั่งซื้อนี้ : {paymentState?.amount} บาท</h1>
           <p className="text-sm">กรุณาชำระเงินก่อน QR CODE หมดอายุ</p>
           <div>{formatTime(time)} นาที</div>
         </div>
